SCMOD-7821: Converter refactored
