+++ caf-logging-logback-access/pom.xml
@@ -27,7 +27,7 @@
         <version>1.2.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>caf-logging-logback-tomcat</artifactId>
+    <artifactId>caf-logging-logback-access</artifactId>
 
     <dependencies>
         <dependency>
@@ -43,5 +43,5 @@
             <artifactId>logback-core</artifactId>
         </dependency>
     </dependencies>
-
+    <name>caf-logging-logback-access</name>
 </project>
